Refactor schedule cog

+++ pyboss/cogs/schedule.py
@@ -7,7 +7,7 @@ import re
 import discord
 from discord.ext import commands
 from discord.ext.commands import guild_only
-from sqlalchemy import insert, select, update
+from sqlalchemy import insert, select, update, delete
 
 from pyboss import CONFIG, STATIC_DIR
 from pyboss.models import CalendarModel, NotebookModel, ScheduleRefModel
@@ -38,18 +38,18 @@ class Schedule:
     with open(STATIC_DIR / "json/matters.json", encoding="utf-8") as wordlist:
         MATTERS = json.load(wordlist)
 
-    def __init__(self, ctx):
+    def __init__(self, ctx, model):
         self.bot = ctx.bot
         self.guild = ctx.guild
         self.author = ctx.author
         self.channel = ctx.channel
+        self.model = model
         self.traced_msg = []
-
-    def is_same_source(self):
-        def predicate(msg: discord.Message):
-            return msg.channel == self.channel and msg.author.id == self.author.id
-
-        return commands.check(predicate)
+        self.subject = None
+        self.response = None
+        self.date = None
+        self.description = None
+        self.starthour, self.endhour = None, None
 
     async def send_question(self, question, check=None, timeout=60):
         msg_question = await self.channel.send(question)
@@ -71,10 +71,12 @@ class Schedule:
 
     def get_data(self, model):
         return database.execute(
-            select(model).where(
+            select(model)
+            .where(
                 model.channel_id == self.channel.id,
                 model.date >= datetime.datetime.now(),
             )
+            .order_by(model.date)
         )
 
     @staticmethod
@@ -82,31 +84,40 @@ class Schedule:
         """
         Update id of agenda or planning message in database
         """
-        ex_message = database.execute(
+        ref_model = database.execute(
             select(ScheduleRefModel).where(
                 ScheduleRefModel.channel_id == message.channel.id,
             )
-        ).scalar_one
-        database.execute(
-            update(ScheduleRefModel)
-            .where(ScheduleRefModel.channel_id == message.channel.id)
-            .values(message_id=message.id)
-        )
-        try:
-            ex_message = await message.channel.fetch_message(ex_message.id)
-        except (discord.NotFound, discord.HTTPException):
-            return
-        await ex_message.delete()
+        ).scalar_one_or_none()
+        if ref_model:
+            database.execute(
+                update(ScheduleRefModel)
+                .where(ScheduleRefModel.channel_id == message.channel.id)
+                .values(message_id=message.id)
+            )
+            try:
+                ex_message = await message.channel.fetch_message(ref_model.message_id)
+            except (discord.NotFound, discord.HTTPException):
+                return
+            await ex_message.delete()
+        else:
+            database.execute(
+                insert(ScheduleRefModel).values(
+                    guild_id=message.guild.id,
+                    channel_id=message.channel.id,
+                    message_id=message.id,
+                )
+            )
 
     async def remove_procedure(self, update_fct=None):
         try:
             await self.send_question(
                 "Quelle matière voulez-vous supprimer?",
                 timeout=60,
-                check=self.is_matter_valid(),
+                check=self.is_subject_valid,
             )
             await self.send_question(
-                "De quel jour? (Ex: JJ/MM)", timeout=60, check=self.is_date_valid()
+                "De quel jour? (Ex: JJ/MM)", timeout=60, check=self.is_date_valid
             )
         except asyncio.TimeoutError:
             await self.channel.send(
@@ -114,108 +125,95 @@ class Schedule:
             )
             raise asyncio.TimeoutError
         else:
+            database.execute(
+                delete(self.model).where(
+                    self.model.subject == self.subject, self.model.date == self.date
+                )
+            )
             if callable(update_fct):
                 await update_fct()
         finally:
             await self.delete_traced_msg()
 
-    def is_matter_valid(self):
+    def is_subject_valid(self, msg: discord.Message):
         """
-        Check if the user write a correct matter and place him in blacklist if not.
+        Check if the user write a correct subject and place him in blacklist if not.
         """
+        if msg.channel != self.channel or msg.author.id != self.author.id:
+            return False
+        member = MemberWrapper(msg.author)
 
-        @self.is_same_source()
-        async def predicate(msg: discord.Message):
-            member = MemberWrapper(msg.author)
+        if subject := self.MATTERS.get(msg.content.upper()):
+            self.subject = subject
+        elif not member.blacklist_date:
+            member.place_in_blacklist()
+            self.subject = msg.content.title()
+        else:
+            seconds = (member.blacklist_date - datetime.datetime.now()).total_seconds()
+            hours, minutes = seconds // 3600, (seconds % 3600) // 60
+            self.response = (
+                f"Vous n'êtes pas autorisé à ajouter un évènement spécial "
+                f"pendant encore {int(hours)}h {int(minutes)}min. "
+            )
+            return False
+        return True
 
-            if matter := self.MATTERS.get(msg.content.upper()):
-                self.matter = matter
-            elif not member.blacklist_date:
-                member.place_in_blacklist()
-                self.matter = msg.content.title()
-            else:
-                seconds = (
-                    member.blacklist_date - datetime.datetime.now()
-                ).total_seconds()
-                hours, minutes = seconds // 3600, (seconds % 3600) // 60
-                await msg.channel.send(
-                    f"Vous n'êtes pas autorisé à ajouter un évènement spécial "
-                    f"pendant encore {int(hours)}h {int(minutes)}min. "
-                )
-                return False
-            return True
-
-        return commands.check(predicate)
-
-    def is_date_valid(self):
-        @self.is_same_source()
-        async def predicate(msg: discord.Message):
-            """
-            Check with a regex the data as user input, can accept several formats
-            """
-            match = REGEX_DATE.match(msg.content)
-            if not match:
-                await msg.channel.send("Votre format de date est invalide.")
-                logger.info(
-                    f"The user {msg.author.name} hasn't wrote correctly the date"
-                )
-                return False
-
-            day, month = map(int, match.groups())
-            try:
-                today = datetime.date.today()
-                year = today.year if today.month > month else today.year + 1
-                date_user = datetime.date(year=year, month=month, day=day)
-            except ValueError:
-                await msg.channel.send("Vous avez mis des valeurs invalides")
-                return False
-
-            if date_user.weekday() < 5:
-                if date_user >= datetime.date.today():
-                    self.date = str(date_user)
-                    return True
-                await msg.channel.send(
-                    f"{day}/{month} est antérieur à la date actuelle"
-                )
-                return False
-            await msg.channel.send("On ne travaille pas le week end!")
+    def is_date_valid(self, msg: discord.Message):
+        """
+        Check with a regex the data as user input, can accept several formats
+        """
+        if msg.channel != self.channel or msg.author.id != self.author.id:
             return False
 
-        return commands.check(predicate)
+        match = REGEX_DATE.match(msg.content)
+        if not match:
+            self.response = "Votre format de date est invalide"
+            logger.info(f"The user {msg.author.name} hasn't wrote correctly the date")
+            return False
 
-    def is_hours_valid(self):
-        @self.is_same_source()
-        async def predicate(msg: discord.Message):
-            try:
-                match = REGEX_HOUR.match(msg.content)
-                starthour, endhour = map(int, match.groups())
-            except (KeyError, ValueError):
-                await msg.channel.send("Votre format de plage horaire est invalide.")
-            else:
-                if not endhour:  # verifying if endhour exists or calculate it
-                    endhour = str(starthour + 1)
-                self.starthour, self.endhour = starthour + "h", endhour + "h"
+        day, month = map(int, match.groups())
+        try:
+            today = datetime.date.today()
+            year = today.year if month >= today.month else today.year + 1
+            date_user = datetime.date(year=year, month=month, day=day)
+        except ValueError:
+            self.response = "Vous avez mis des valeurs invalides"
+            return False
+
+        if date_user.weekday() < 5:
+            if date_user >= datetime.date.today():
+                self.date = str(date_user)
                 return True
+            self.response = f"{day}/{month} est antérieur à la date actuelle"
+            return False
+        self.response = "On ne travaille pas le week end!"
+        return False
 
-        return commands.check(predicate)
-
-    def is_description_valid(self):
-        @self.is_same_source()
-        async def predicate(msg: discord.Message):
-            if msg.content.strip().upper() != "NON":
-                self.description = msg.content
+    def is_hours_valid(self, msg: discord.Message):
+        if msg.channel != self.channel or msg.author.id != self.author.id:
+            return False
+        try:
+            match = REGEX_HOUR.match(msg.content)
+            starthour = int(match.group(1))
+        except (KeyError, ValueError):
+            self.response = "Votre format de plage horaire est invalide."
+        else:
+            if not (endhour := match.group(2)):
+                endhour = str(starthour + 1)
+            self.starthour, self.endhour = f"{starthour}h", f"{endhour}h"
             return True
 
-        return commands.check(predicate)
+    def is_description_valid(self, msg: discord.Message):
+        if msg.channel != self.channel or msg.author.id != self.author.id:
+            return False
+        if msg.content.strip().upper() != "NON":
+            self.description = msg.content
+        return True
 
 
 class Calendar(Schedule):
     def __init__(self, ctx):
-        super().__init__(ctx)
-        self.subject = None
-        self.date = None
-        self.starthour, self.endhour = None, None
-        self.description = None
+        super().__init__(ctx, CalendarModel)
 
     async def rules(self):
         """
@@ -229,30 +227,32 @@ class Calendar(Schedule):
     async def new(self):
         try:
             await self.send_question(
-                "Quelle matière voulez-vous ajouter ?", self.is_matter_valid()
+                "Quelle matière voulez-vous ajouter ?", self.is_subject_valid
             )
             await self.send_question(
                 "A quelle date le cours aura lieu ? (Ex : JJ/MM)",
-                self.is_date_valid(),
+                self.is_date_valid,
             )
             await self.send_question(
-                "A quelle heure ? (Ex: 10h; 10h/12h…)", self.is_hours_valid()
+                "A quelle heure ? (Ex: 10h; 10h/12h…)", self.is_hours_valid
             )
             await self.send_question(
                 "Voulez-vous ajouter des infos supplémentaires? "
                 "('votre contenu' ou 'non')",
-                self.is_description_valid(),
+                self.is_description_valid,
                 timeout=120,
             )
         except asyncio.TimeoutError:
             await self.channel.send(
                 f"{self.author.mention} Vous avez mis trop de temps à répondre"
             )
+            if self.response:
+                await self.channel.send(f"Détails: \n{self.response}")
         else:
             database.execute(
                 insert(CalendarModel).values(
                     guild_id=self.guild.id,
-                    channel=self.channel.name,
+                    channel_id=self.channel.id,
                     subject=self.subject,
                     description=self.description,
                     date=self.date,
@@ -271,18 +271,15 @@ class Calendar(Schedule):
         next_date, message = None, ""
         for row in self.get_data(CalendarModel):
             date = row[0].date
-
             if date != next_date:
                 next_date = date
                 day = DAYS[date.weekday()]
                 month = MONTHS[int(f"{date:%m}") - 1]
                 message += f"\n__Pour le **{day}** {date:%d} {month}:__\n"
 
-            description = (
-                f"*({row[0].description})*" if bool(row[0].description) else ""
-            )
+            description = f"*({row[0].description})*" if row[0].description else ""
             message += (
-                f"- \t {row[0].matter}: "
+                f"- \t {row[0].subject}: "
                 f"**{row[0].starthour}**-**{row[0].endhour}** {description}\n"
             )
 
@@ -296,10 +293,7 @@ class Calendar(Schedule):
 
 class Notebook(Schedule):
     def __init__(self, ctx):
-        super().__init__(ctx)
-        self.subject = None
-        self.date = None
-        self.description = None
+        super().__init__(ctx, NotebookModel)
 
     async def rules(self):
         """
@@ -313,26 +307,28 @@ class Notebook(Schedule):
     async def new(self):
         try:
             await self.send_question(
-                "Quelle matière voulez-vous ajouter ?", self.is_matter_valid()
+                "Quelle matière voulez-vous ajouter ?", self.is_subject_valid
             )
             await self.send_question(
                 "Quelle est la date limite du devoir ? (Ex : JJ/MM)",
-                self.is_date_valid(),
+                self.is_date_valid,
             )
             await self.send_question(
                 "Indiquez quels sont les devoirs à faire (obligatoire):",
-                self.is_description_valid(),
+                self.is_description_valid,
                 timeout=120,
             )
         except asyncio.TimeoutError:
             await self.channel.send(
                 f"{self.author.mention} Vous avez mis trop de temps à répondre"
             )
+            if self.response:
+                await self.channel.send(f"Détails: \n{self.response}")
         else:
             database.execute(
                 insert(NotebookModel).values(
                     guild_id=self.guild.id,
-                    channel=self.channel.name,
+                    channel_id=self.channel.id,
                     subject=self.subject,
                     description=self.description,
                     date=self.date,
